LayoutTests/inspector/dom-debugger/url-breakpoints-dom-expected.txt

Tests for URL breakpoints.


== Running test suite: URLBreakpoint
-- Running test case: URLBreakpoint.ShouldNotPauseOnDisabledBreakpoint.DOM
Adding "text:data" URL Breakpoint...
Disabling breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: URLBreakpoint.ShouldNotPauseOnNotContainingURL.DOM
Adding "text:nonexistant-url" URL Breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause for breakpoint with different URL.
-- Running test teardown.

-- Running test case: URLBreakpoint.RemoveBreakpoint.DOM
Adding "text:data" URL Breakpoint...
Removing breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: URLBreakpoint.RemoveDisabledBreakpoint.DOM
Adding "text:data" URL Breakpoint...
Disabling breakpoint...
Removing breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.
